… v2/event/event_marshal.go
… v2/event/event_marshal_benchmark_test.go
… v2/event/event_marshal_test.go
… v2/event/event_unmarshal_test.go