Fix #609

This PR is definitely incomplete and the code has to be reorganized, well documented and optimized. If, even after optimizing it, the speedup is not significant, i propose to drop it, since, even with some reorganization of the code, it adds a lot of complexity.

Preliminary results:

```
name                                                                               old time/op    new time/op    delta
Unmarshal/struct_data_v0.3-8                                                         30.5µs ± 0%    27.9µs ± 0%   ~     (p=1.000 n=1+1)
Unmarshal/nil_data_v1.0-8                                                            30.7µs ± 0%    28.3µs ± 0%   ~     (p=1.000 n=1+1)
Unmarshal/string_data_v1.0-8                                                         33.4µs ± 0%    29.8µs ± 0%   ~     (p=1.000 n=1+1)
Unmarshal/base64_json_encoded_data_v1.0-8                                            18.0µs ± 0%    18.1µs ± 0%   ~     (p=1.000 n=1+1)
Unmarshal/base64_xml_encoded_data_v1.0-8                                             18.7µs ± 0%    18.6µs ± 0%   ~     (p=1.000 n=1+1)
Unmarshal/string_data_v0.3-8                                                         30.9µs ± 0%    29.2µs ± 0%   ~     (p=1.000 n=1+1)
Unmarshal/nil_data_v0.3-8                                                            28.5µs ± 0%    26.9µs ± 0%   ~     (p=1.000 n=1+1)
Unmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v0.3-8    30.3µs ± 0%    28.0µs ± 0%   ~     (p=1.000 n=1+1)
Unmarshal/struct_data_v1.0-8                                                         32.6µs ± 0%    31.2µs ± 0%   ~     (p=1.000 n=1+1)
Unmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v1.0-8    33.0µs ± 0%    29.4µs ± 0%   ~     (p=1.000 n=1+1)

name                                                                               old alloc/op   new alloc/op   delta
Unmarshal/struct_data_v0.3-8                                                         9.68kB ± 0%   11.62kB ± 0%   ~     (p=1.000 n=1+1)
Unmarshal/nil_data_v1.0-8                                                            8.09kB ± 0%   11.34kB ± 0%   ~     (p=1.000 n=1+1)
Unmarshal/string_data_v1.0-8                                                         9.62kB ± 0%   11.69kB ± 0%   ~     (p=1.000 n=1+1)
Unmarshal/base64_json_encoded_data_v1.0-8                                            5.37kB ± 0%    7.87kB ± 0%   ~     (p=1.000 n=1+1)
Unmarshal/base64_xml_encoded_data_v1.0-8                                             5.42kB ± 0%    7.92kB ± 0%   ~     (p=1.000 n=1+1)
Unmarshal/string_data_v0.3-8                                                         9.63kB ± 0%   11.68kB ± 0%   ~     (p=1.000 n=1+1)
Unmarshal/nil_data_v0.3-8                                                            8.09kB ± 0%   11.33kB ± 0%   ~     (p=1.000 n=1+1)
Unmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v0.3-8    9.68kB ± 0%   11.62kB ± 0%   ~     (p=1.000 n=1+1)
Unmarshal/struct_data_v1.0-8                                                         9.66kB ± 0%   11.62kB ± 0%   ~     (p=1.000 n=1+1)
Unmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v1.0-8    9.66kB ± 0%   11.62kB ± 0%   ~     (p=1.000 n=1+1)

name                                                                               old allocs/op  new allocs/op  delta
Unmarshal/struct_data_v0.3-8                                                            157 ± 0%       289 ± 0%   ~     (p=1.000 n=1+1)
Unmarshal/nil_data_v1.0-8                                                               148 ± 0%       275 ± 0%   ~     (p=1.000 n=1+1)
Unmarshal/string_data_v1.0-8                                                            154 ± 0%       291 ± 0%   ~     (p=1.000 n=1+1)
Unmarshal/base64_json_encoded_data_v1.0-8                                               100 ± 0%       177 ± 0%   ~     (p=1.000 n=1+1)
Unmarshal/base64_xml_encoded_data_v1.0-8                                                100 ± 0%       177 ± 0%   ~     (p=1.000 n=1+1)
Unmarshal/string_data_v0.3-8                                                            154 ± 0%       291 ± 0%   ~     (p=1.000 n=1+1)
Unmarshal/nil_data_v0.3-8                                                               148 ± 0%       275 ± 0%   ~     (p=1.000 n=1+1)
Unmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v0.3-8       157 ± 0%       289 ± 0%   ~     (p=1.000 n=1+1)
Unmarshal/struct_data_v1.0-8                                                            157 ± 0%       289 ± 0%   ~     (p=1.000 n=1+1)
Unmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v1.0-8       157 ± 0%       289 ± 0%   ~     (p=1.000 n=1+1)
```