@slinkydeveloper 2020-11-23
I reorganized the benchmark to make it reproducible (when it was serializing the input, there was no guarantee on the ordering of json fields):

```
name                                                                               old time/op    new time/op    delta
Unmarshal/nil_data_v0.3-8                                                            24.2µs ± 1%    21.8µs ± 1%   -9.81%  (p=0.008 n=5+5)
Unmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v1.0-8    28.6µs ± 1%    29.3µs ± 1%   +2.54%  (p=0.008 n=5+5)
Unmarshal/nil_data_v1.0-8                                                            26.2µs ± 0%    24.1µs ± 2%   -8.31%  (p=0.008 n=5+5)
Unmarshal/struct_data_v0.3-8                                                         27.0µs ± 0%    23.0µs ± 1%  -14.60%  (p=0.008 n=5+5)
Unmarshal/string_data_v0.3-8                                                         26.6µs ± 0%    23.0µs ± 1%  -13.70%  (p=0.008 n=5+5)
Unmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v0.3-8    26.9µs ± 1%    27.6µs ± 1%   +2.47%  (p=0.008 n=5+5)
Unmarshal/struct_data_v1.0-8                                                         29.0µs ± 1%    25.0µs ± 1%  -13.87%  (p=0.008 n=5+5)
Unmarshal/string_data_v1.0-8                                                         28.6µs ± 0%    25.0µs ± 1%  -12.79%  (p=0.008 n=5+5)
Unmarshal/base64_json_encoded_data_v1.0-8                                            15.9µs ± 0%    15.2µs ± 1%   -4.80%  (p=0.008 n=5+5)
Unmarshal/base64_xml_encoded_data_v1.0-8                                             16.7µs ± 1%    15.7µs ± 1%   -5.97%  (p=0.008 n=5+5)

name                                                                               old alloc/op   new alloc/op   delta
Unmarshal/nil_data_v0.3-8                                                            8.09kB ± 0%    7.81kB ± 0%     ~     (p=0.079 n=4+5)
Unmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v1.0-8    9.66kB ± 0%   11.48kB ± 0%  +18.80%  (p=0.008 n=5+5)
Unmarshal/nil_data_v1.0-8                                                            8.09kB ± 0%    7.81kB ± 0%   -3.44%  (p=0.008 n=5+5)
Unmarshal/struct_data_v0.3-8                                                         9.67kB ± 0%    8.05kB ± 0%     ~     (p=0.079 n=4+5)
Unmarshal/string_data_v0.3-8                                                         9.62kB ± 0%    8.11kB ± 0%  -15.65%  (p=0.008 n=5+5)
Unmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v0.3-8    9.66kB ± 0%   11.47kB ± 0%  +18.79%  (p=0.008 n=5+5)
Unmarshal/struct_data_v1.0-8                                                         9.67kB ± 0%    8.05kB ± 0%  -16.73%  (p=0.008 n=5+5)
Unmarshal/string_data_v1.0-8                                                         9.62kB ± 0%    8.12kB ± 0%     ~     (p=0.079 n=4+5)
Unmarshal/base64_json_encoded_data_v1.0-8                                            5.37kB ± 0%    6.07kB ± 0%  +12.98%  (p=0.008 n=5+5)
Unmarshal/base64_xml_encoded_data_v1.0-8                                             5.42kB ± 0%    6.08kB ± 0%  +12.28%  (p=0.008 n=5+5)

name                                                                               old allocs/op  new allocs/op  delta
Unmarshal/nil_data_v0.3-8                                                               148 ± 0%       242 ± 0%  +63.51%  (p=0.008 n=5+5)
Unmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v1.0-8       157 ± 0%       303 ± 0%  +92.99%  (p=0.008 n=5+5)
Unmarshal/nil_data_v1.0-8                                                               148 ± 0%       242 ± 0%  +63.51%  (p=0.008 n=5+5)
Unmarshal/struct_data_v0.3-8                                                            157 ± 0%       255 ± 0%  +62.42%  (p=0.008 n=5+5)
Unmarshal/string_data_v0.3-8                                                            154 ± 0%       257 ± 0%  +66.88%  (p=0.008 n=5+5)
Unmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v0.3-8       157 ± 0%       303 ± 0%  +92.99%  (p=0.008 n=5+5)
Unmarshal/struct_data_v1.0-8                                                            157 ± 0%       255 ± 0%  +62.42%  (p=0.008 n=5+5)
Unmarshal/string_data_v1.0-8                                                            154 ± 0%       257 ± 0%  +66.88%  (p=0.008 n=5+5)
Unmarshal/base64_json_encoded_data_v1.0-8                                               100 ± 0%       159 ± 0%  +59.00%  (p=0.008 n=5+5)
Unmarshal/base64_xml_encoded_data_v1.0-8                                                100 ± 0%       159 ± 0%  +59.00%  (p=0.008 n=5+5)
```

There is an interesting delta in the cpu time when fields are ordered in a way that always hits the best case, when it hits the worst case the delta is +2% higher. I'm gonna continue to investigate, in order to reduce that +2% higher to 0%