@slinkydeveloper 2020-11-24
Closed in favour of https://github.com/cloudevents/sdk-go/pull/618